Refactoring: reuse
